separate util.h into header and implementation files
